move boundaries.py to top level and update imports
