+++ pycbc/boundaries.py
